requirements.txt
PySide6>=6.5.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
pyserial>=3.5
